requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/dev.in
#
aspy.yaml==0.3.0          # via pre-commit
cached-property==1.3.0    # via pre-commit
click==6.7                # via pip-tools
django-localflavor==1.4.1
django==1.11
first==2.0.1              # via pip-tools
nodeenv==1.1.2            # via pre-commit
pip-tools==1.9.0
pre-commit==0.13.6
py==1.4.33                # via pytest
pycodestyle==2.3.1
pytest==3.0.7
pytz==2017.2
pyyaml==3.12              # via aspy.yaml, pre-commit
six==1.10.0               # via pip-tools, pre-commit
virtualenv==15.1.0        # via pre-commit
